Add function overloading
Multiple functions can be defined with the same name. To determine which
function is used in an expression, the compiler simply tries each one,
starting with the most recently defined, until it finds one that creates
a well-typed expression. This means there can be multiple functions with
identical types - the last one will be used. This allows a kind of
"function shadowing".

+++ src/compiler/common.rs
@@ -1,3 +1,3 @@
 pub type Location = usize;
 
-pub type Id = u64;
+pub type Id = usize;
